+++ srcs/push_swap/init_b_to_a.c
@@ -6,7 +6,7 @@
 /*   By: rjesus-d <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/21 11:05:07 by rjesus-d          #+#    #+#             */
-/*   Updated: 2025/01/27 17:41:30 by rjesus-d         ###   ########.fr       */
+/*   Updated: 2025/01/28 17:42:00 by rjesus-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,11 +16,11 @@ static void	set_target_b(t_stack_node *a, t_stack_node *b)
 {
 	t_stack_node	*current_a;
 	t_stack_node	*target_node;
-	int				best_match_index;
+	long			best_match_index;
 
 	while (b)
 	{
-		best_match_index = INT_MAX;
+		best_match_index = LONG_MAX;
 		current_a = a;
 		while (current_a)
 		{
@@ -31,7 +31,7 @@ static void	set_target_b(t_stack_node *a, t_stack_node *b)
 			}
 			current_a = current_a->next;
 		}
-		if (best_match_index == INT_MAX)
+		if (best_match_index == LONG_MAX)
 			b->target_node = find_min(a);
 		else
 			b->target_node = target_node;
